Store Travis token in Google Key Management Service (KMS) (#4)
Google App Engine does not have a defined mechanism to securely set environment variables in the deployment environment, as it only allows installing raw env vars in `app.yaml`.

Per the recommendation from https://stackoverflow.com/a/56107715, we safely workaround this by using Google Cloud KMS. We created a key in the Google Cloud console that was then used to run `gcloud kms encrypt` against the `travis token` from the account `pantsbuild-ci-bot`. Then, the server uses the KMS client library to decrypt this token and authenticate against Travis.

We must use base64 encryption on the already encrypted Travis token to make the value ASCII, so that we can put it safely in `app.yaml`.