Stop caring what specific shard the request comes from (#14)
We want to start remoting the lint shard now that we have V2 isort. To do this, we simply stop caring what the shard number is and allow any shard from an authenticated build to generate a token. 

This is slightly less secure, but not by enough to matter. Someone could just as easily grab a token through the authorized shards. 

The benefit of this change is that we decouple this repository from any changes to Pants CI remote execution.